converter/storage: move disk metadata helpers to storage/disk.go
Like the volume source helpers moved in the previous commit, these
functions only serve the disk conversion loop. Moving them completes
the prerequisite for extracting the loop into a DiskConfigurator.

Move disk naming, block size detection, IO thread assignment, error
policy, and reservation helpers from converter.go to storage/disk.go.

Tests for NewDeviceNamer and makeDeviceName move to
storage/disk_test.go since they access unexported struct fields.
