+++ pkg/virt-launcher/virtwrap/converter/BUILD.bazel
@@ -13,10 +13,6 @@ go_library(
     deps = [
         "//pkg/handler-launcher-com/cmd/v1:go_default_library",
         "//pkg/os/disk:go_default_library",
-        "//pkg/pointer:go_default_library",
-        "//pkg/safepath:go_default_library",
-        "//pkg/storage/reservation:go_default_library",
-        "//pkg/storage/types:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/hardware:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
@@ -30,13 +26,11 @@ go_library(
         "//pkg/virt-launcher/virtwrap/converter/types:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter/vcpu:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter/virtio:go_default_library",
-        "//pkg/virt-launcher/virtwrap/device:go_default_library",
         "//pkg/virt-launcher/virtwrap/disksource:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//staging/src/kubevirt.io/client-go/precond:go_default_library",
         "//vendor/go.uber.org/mock/gomock:go_default_library",
-        "//vendor/golang.org/x/sys/unix:go_default_library",
         "//vendor/k8s.io/utils/ptr:go_default_library",
     ],
 )
