+++ pkg/virt-launcher/virtwrap/converter/converter_test.go
@@ -260,7 +260,7 @@ var _ = Describe("Converter", func() {
 				},
 			}
 			apiDisk := api.Disk{}
-			devicePerBus := map[string]deviceNamer{}
+			devicePerBus := map[string]storage.DeviceNamer{}
 			numQueues := uint(2)
 			volumeStatusMap := make(map[string]v1.VolumeStatus)
 			volumeStatusMap["myvolume"] = v1.VolumeStatus{
@@ -273,7 +273,7 @@ var _ = Describe("Converter", func() {
 					},
 				},
 			}
-			Convert_v1_Disk_To_api_Disk(context, &v1Disk, &apiDisk, devicePerBus, &numQueues, volumeStatusMap)
+			storage.Convert_v1_Disk_To_api_Disk(context, &v1Disk, &apiDisk, devicePerBus, &numQueues, volumeStatusMap)
 			Expect(apiDisk.Capacity).ToNot(BeNil())
 			Expect(*apiDisk.Capacity).To(Equal(expected))
 		},
@@ -288,11 +288,11 @@ var _ = Describe("Converter", func() {
 				DiskDevice: diskDevice,
 			}
 			apiDisk := api.Disk{}
-			devicePerBus := map[string]deviceNamer{}
+			devicePerBus := map[string]storage.DeviceNamer{}
 			numQueues := uint(2)
 			volumeStatusMap := make(map[string]v1.VolumeStatus)
 			volumeStatusMap["myvolume"] = v1.VolumeStatus{}
-			Convert_v1_Disk_To_api_Disk(context, &v1Disk, &apiDisk, devicePerBus, &numQueues, volumeStatusMap)
+			storage.Convert_v1_Disk_To_api_Disk(context, &v1Disk, &apiDisk, devicePerBus, &numQueues, volumeStatusMap)
 			Expect(apiDisk.Address).ToNot(BeNil())
 			Expect(apiDisk.Address.Bus).To(Equal("0"))
 			Expect(apiDisk.Address.Controller).To(Equal("0"))
@@ -397,7 +397,7 @@ var _ = Describe("Converter", func() {
 				},
 			}
 			libvirtDisk := &api.Disk{}
-			err := Convert_v1_BlockSize_To_api_BlockIO(kubevirtDisk, libvirtDisk, arch)
+			err := storage.Convert_v1_BlockSize_To_api_BlockIO(kubevirtDisk, libvirtDisk, arch)
 			if shouldSucceed {
 				Expect(err).ToNot(HaveOccurred())
 				expectedXML := fmt.Sprintf(`<Disk device="" type="">
@@ -1767,7 +1767,7 @@ var _ = Describe("Converter", func() {
 					},
 				}
 				apiDisk := api.Disk{Source: api.DiskSource{File: "/"}}
-				Expect(Convert_v1_BlockSize_To_api_BlockIO(&v1Disk, &apiDisk, amd64)).To(Succeed())
+				Expect(storage.Convert_v1_BlockSize_To_api_BlockIO(&v1Disk, &apiDisk, amd64)).To(Succeed())
 
 				blockIO := apiDisk.BlockIO
 				Expect(blockIO.LogicalBlockSize).To(Equal(blockIO.PhysicalBlockSize))
@@ -1789,7 +1789,7 @@ var _ = Describe("Converter", func() {
 					},
 				}
 				apiDisk := api.Disk{Source: api.DiskSource{}}
-				Expect(Convert_v1_BlockSize_To_api_BlockIO(&v1Disk, &apiDisk, amd64)).To(MatchError(ContainSubstring(blockIoConfigErrorMessage)))
+				Expect(storage.Convert_v1_BlockSize_To_api_BlockIO(&v1Disk, &apiDisk, amd64)).To(MatchError(ContainSubstring(blockIoConfigErrorMessage)))
 			})
 
 			It("Should fail block size detection for a nil domain disk", func() {
@@ -1800,7 +1800,7 @@ var _ = Describe("Converter", func() {
 						MatchVolume: &v1.FeatureState{Enabled: pointer.P(true)},
 					},
 				}
-				Expect(Convert_v1_BlockSize_To_api_BlockIO(&v1Disk, nil, amd64)).To(MatchError(ContainSubstring(nilDiskErrorMessage)))
+				Expect(storage.Convert_v1_BlockSize_To_api_BlockIO(&v1Disk, nil, amd64)).To(MatchError(ContainSubstring(nilDiskErrorMessage)))
 			})
 		})
 	})
@@ -2908,9 +2908,9 @@ var _ = Describe("Converter", func() {
 				},
 			}
 			apiDisk := api.Disk{}
-			devicePerBus := map[string]deviceNamer{}
+			devicePerBus := map[string]storage.DeviceNamer{}
 			numQueues := uint(2)
-			Convert_v1_Disk_To_api_Disk(context, &v1Disk, &apiDisk, devicePerBus, &numQueues, make(map[string]v1.VolumeStatus))
+			storage.Convert_v1_Disk_To_api_Disk(context, &v1Disk, &apiDisk, devicePerBus, &numQueues, make(map[string]v1.VolumeStatus))
 			Expect(apiDisk.Device).To(Equal("disk"), "expected disk device to be defined")
 			Expect(*(apiDisk.Driver.Queues)).To(Equal(expectedQueues), "expected queues to be 2")
 		})
@@ -2922,8 +2922,8 @@ var _ = Describe("Converter", func() {
 				},
 			}
 			apiDisk := api.Disk{}
-			devicePerBus := map[string]deviceNamer{}
-			Expect(Convert_v1_Disk_To_api_Disk(context, &v1Disk, &apiDisk, devicePerBus, nil, make(map[string]v1.VolumeStatus))).
+			devicePerBus := map[string]storage.DeviceNamer{}
+			Expect(storage.Convert_v1_Disk_To_api_Disk(context, &v1Disk, &apiDisk, devicePerBus, nil, make(map[string]v1.VolumeStatus))).
 				To(Succeed())
 			Expect(apiDisk.Device).To(Equal("disk"), "expected disk device to be defined")
 			Expect(apiDisk.Driver.Queues).To(BeNil(), "expected no queues to be requested")
@@ -3139,164 +3139,6 @@ var _ = Describe("Converter", func() {
 					},
 				}),
 		)
-	})
-
-	Describe("newDeviceNamer", func() {
-		DescribeTable("should correctly build prefix map for different disk types",
-			func(volumeStatuses []v1.VolumeStatus, disks []v1.Disk, expectedPrefixes []string, expectedMappings map[string]map[string]string) {
-				prefixMap := newDeviceNamer(volumeStatuses, disks)
-
-				// Verify expected prefixes exist
-				Expect(prefixMap).To(HaveLen(len(expectedPrefixes)))
-				for _, prefix := range expectedPrefixes {
-					Expect(prefixMap).To(HaveKey(prefix))
-				}
-
-				// Verify mappings for each prefix
-				for prefix, mappings := range expectedMappings {
-					namer := prefixMap[prefix]
-					for diskName, target := range mappings {
-						Expect(namer.existingNameMap).To(HaveKeyWithValue(diskName, target))
-						Expect(namer.usedDeviceMap).To(HaveKeyWithValue(target, diskName))
-					}
-				}
-			},
-			Entry("Disk with virtio bus",
-				[]v1.VolumeStatus{{Name: "mydisk", Target: "vda"}},
-				[]v1.Disk{{Name: "mydisk", DiskDevice: v1.DiskDevice{Disk: &v1.DiskTarget{Bus: v1.DiskBusVirtio}}}},
-				[]string{"vd"},
-				map[string]map[string]string{"vd": {"mydisk": "vda"}},
-			),
-			Entry("Disk with SATA bus",
-				[]v1.VolumeStatus{{Name: "mydisk", Target: "sda"}},
-				[]v1.Disk{{Name: "mydisk", DiskDevice: v1.DiskDevice{Disk: &v1.DiskTarget{Bus: v1.DiskBusSATA}}}},
-				[]string{"sd"},
-				map[string]map[string]string{"sd": {"mydisk": "sda"}},
-			),
-			Entry("LUN with SCSI bus",
-				[]v1.VolumeStatus{{Name: "mylun", Target: "sda"}},
-				[]v1.Disk{{Name: "mylun", DiskDevice: v1.DiskDevice{LUN: &v1.LunTarget{Bus: v1.DiskBusSCSI}}}},
-				[]string{"sd"},
-				map[string]map[string]string{"sd": {"mylun": "sda"}},
-			),
-			Entry("CDRom with SATA bus",
-				[]v1.VolumeStatus{{Name: "mycdrom", Target: "sda"}},
-				[]v1.Disk{{Name: "mycdrom", DiskDevice: v1.DiskDevice{CDRom: &v1.CDRomTarget{Bus: v1.DiskBusSATA}}}},
-				[]string{"sd"},
-				map[string]map[string]string{"sd": {"mycdrom": "sda"}},
-			),
-			Entry("Mixed devices with different buses",
-				[]v1.VolumeStatus{
-					{Name: "disk1", Target: "vda"},
-					{Name: "lun1", Target: "sda"},
-					{Name: "cdrom1", Target: "sdb"},
-				},
-				[]v1.Disk{
-					{Name: "disk1", DiskDevice: v1.DiskDevice{Disk: &v1.DiskTarget{Bus: v1.DiskBusVirtio}}},
-					{Name: "lun1", DiskDevice: v1.DiskDevice{LUN: &v1.LunTarget{Bus: v1.DiskBusSCSI}}},
-					{Name: "cdrom1", DiskDevice: v1.DiskDevice{CDRom: &v1.CDRomTarget{Bus: v1.DiskBusSATA}}},
-				},
-				[]string{"vd", "sd"},
-				map[string]map[string]string{
-					"vd": {"disk1": "vda"},
-					"sd": {"lun1": "sda", "cdrom1": "sdb"},
-				},
-			),
-			Entry("Multiple disks sharing same prefix",
-				[]v1.VolumeStatus{
-					{Name: "disk1", Target: "vda"},
-					{Name: "disk2", Target: "vdb"},
-				},
-				[]v1.Disk{
-					{Name: "disk1", DiskDevice: v1.DiskDevice{Disk: &v1.DiskTarget{Bus: v1.DiskBusVirtio}}},
-					{Name: "disk2", DiskDevice: v1.DiskDevice{Disk: &v1.DiskTarget{Bus: v1.DiskBusVirtio}}},
-				},
-				[]string{"vd"},
-				map[string]map[string]string{"vd": {"disk1": "vda", "disk2": "vdb"}},
-			),
-		)
-
-		It("should return empty map for empty inputs", func() {
-			prefixMap := newDeviceNamer([]v1.VolumeStatus{}, []v1.Disk{})
-			Expect(prefixMap).To(BeEmpty())
-		})
-
-		It("should not populate maps when volume status has no target", func() {
-			volumeStatuses := []v1.VolumeStatus{{Name: "mydisk", Target: ""}}
-			disks := []v1.Disk{{Name: "mydisk", DiskDevice: v1.DiskDevice{Disk: &v1.DiskTarget{Bus: v1.DiskBusVirtio}}}}
-
-			prefixMap := newDeviceNamer(volumeStatuses, disks)
-
-			Expect(prefixMap).To(HaveKey("vd"))
-			namer := prefixMap["vd"]
-			Expect(namer.existingNameMap).To(BeEmpty())
-			Expect(namer.usedDeviceMap).To(BeEmpty())
-		})
-
-		It("should skip disks with no Disk, LUN, or CDRom defined", func() {
-			volumeStatuses := []v1.VolumeStatus{{Name: "unknown", Target: "vda"}}
-			disks := []v1.Disk{{Name: "unknown", DiskDevice: v1.DiskDevice{}}}
-
-			prefixMap := newDeviceNamer(volumeStatuses, disks)
-
-			Expect(prefixMap).To(BeEmpty())
-		})
-
-		It("should handle disk without matching volume status", func() {
-			volumeStatuses := []v1.VolumeStatus{} // no volume status
-			disks := []v1.Disk{{Name: "mydisk", DiskDevice: v1.DiskDevice{Disk: &v1.DiskTarget{Bus: v1.DiskBusVirtio}}}}
-
-			prefixMap := newDeviceNamer(volumeStatuses, disks)
-
-			Expect(prefixMap).To(HaveKey("vd"))
-			namer := prefixMap["vd"]
-			// Prefix entry exists but no mappings since there's no volume status with target
-			Expect(namer.existingNameMap).To(BeEmpty())
-			Expect(namer.usedDeviceMap).To(BeEmpty())
-		})
-
-		It("should allow new disk to fill gap after hotplug detach with mixed disk types", func() {
-			By("Simulating post-detach state: boot=sda, detached=sdb(missing), cdrom=sdc, cloudinit=sdd")
-			// This scenario: SCSI boot disk at sda, hotplugged SCSI disk was detached from sdb,
-			// SATA cdrom at sdc, SATA cloudinit at sdd. All share the 'sd' prefix.
-			volumeStatuses := []v1.VolumeStatus{
-				{Name: "boot", Target: "sda"},
-				// sdb is missing - was detached
-				{Name: "cdrom", Target: "sdc"},
-				{Name: "cloudinit", Target: "sdd"},
-			}
-			disks := []v1.Disk{
-				{Name: "boot", DiskDevice: v1.DiskDevice{Disk: &v1.DiskTarget{Bus: v1.DiskBusSCSI}}},
-				{Name: "cdrom", DiskDevice: v1.DiskDevice{CDRom: &v1.CDRomTarget{Bus: v1.DiskBusSATA}}},
-				{Name: "cloudinit", DiskDevice: v1.DiskDevice{CDRom: &v1.CDRomTarget{Bus: v1.DiskBusSATA}}},
-			}
-
-			prefixMap := newDeviceNamer(volumeStatuses, disks)
-
-			By("Verifying the prefix map correctly tracks all used targets")
-			Expect(prefixMap).To(HaveKey("sd"))
-			namer := prefixMap["sd"]
-			Expect(namer.usedDeviceMap).To(HaveKeyWithValue("sda", "boot"))
-			Expect(namer.usedDeviceMap).To(HaveKeyWithValue("sdc", "cdrom"))
-			Expect(namer.usedDeviceMap).To(HaveKeyWithValue("sdd", "cloudinit"))
-			Expect(namer.usedDeviceMap).NotTo(HaveKey("sdb")) // sdb should be available
-
-			By("Simulating domain conversion: calling makeDeviceName for each disk in order")
-			// This mirrors what Convert_v1_VirtualMachineInstance_To_api_Domain does
-			bootTarget, _ := makeDeviceName("boot", v1.DiskBusSCSI, prefixMap)
-			Expect(bootTarget).To(Equal("sda"), "boot disk should keep its existing target")
-
-			cdromTarget, _ := makeDeviceName("cdrom", v1.DiskBusSATA, prefixMap)
-			Expect(cdromTarget).To(Equal("sdc"), "cdrom should keep its existing target, not get reassigned")
-
-			cloudinitTarget, _ := makeDeviceName("cloudinit", v1.DiskBusSATA, prefixMap)
-			Expect(cloudinitTarget).To(Equal("sdd"), "cloudinit should keep its existing target, not get reassigned")
-
-			By("Adding a new hotplug SCSI disk - it should get sdb (the gap)")
-			newDiskName, index := makeDeviceName("newhotplug", v1.DiskBusSCSI, prefixMap)
-			Expect(newDiskName).To(Equal("sdb"), "new disk should fill the gap at sdb, not collide with sdc/sdd")
-			Expect(index).To(Equal(1))
-		})
 	})
 
 	Context("Correctly handle IsolateEmulatorThread with dedicated cpus", func() {
@@ -4606,73 +4448,19 @@ var _ = Describe("Converter", func() {
 
 var _ = Describe("disk device naming", func() {
 	It("format device name should return correct value", func() {
-		res := FormatDeviceName("sd", 0)
+		res := storage.FormatDeviceName("sd", 0)
 		Expect(res).To(Equal("sda"))
-		res = FormatDeviceName("sd", 1)
+		res = storage.FormatDeviceName("sd", 1)
 		Expect(res).To(Equal("sdb"))
 		// 25 is z 26 starting at 0
-		res = FormatDeviceName("sd", 25)
+		res = storage.FormatDeviceName("sd", 25)
 		Expect(res).To(Equal("sdz"))
-		res = FormatDeviceName("sd", 26*2-1)
+		res = storage.FormatDeviceName("sd", 26*2-1)
 		Expect(res).To(Equal("sdaz"))
-		res = FormatDeviceName("sd", 26*26-1)
+		res = storage.FormatDeviceName("sd", 26*26-1)
 		Expect(res).To(Equal("sdyz"))
 	})
 
-	It("makeDeviceName should generate proper name", func() {
-		prefixMap := make(map[string]deviceNamer)
-		res, index := makeDeviceName("test1", v1.VirtIO, prefixMap)
-		Expect(res).To(Equal("vda"))
-		Expect(index).To(Equal(0))
-		for i := 2; i < 10; i++ {
-			makeDeviceName(fmt.Sprintf("test%d", i), v1.VirtIO, prefixMap)
-		}
-		prefix := getPrefixFromBus(v1.VirtIO)
-		delete(prefixMap[prefix].usedDeviceMap, "vdd")
-		By("Verifying next value is vdd")
-		res, index = makeDeviceName("something", v1.VirtIO, prefixMap)
-		Expect(index).To(Equal(3))
-		Expect(res).To(Equal("vdd"))
-		res, index = makeDeviceName("something_else", v1.VirtIO, prefixMap)
-		Expect(res).To(Equal("vdj"))
-		Expect(index).To(Equal(9))
-		By("verifying existing returns correct value")
-		res, index = makeDeviceName("something", v1.VirtIO, prefixMap)
-		Expect(res).To(Equal("vdd"))
-		Expect(index).To(Equal(3))
-		By("Verifying a new bus returns from start")
-		res, index = makeDeviceName("something", "scsi", prefixMap)
-		Expect(res).To(Equal("sda"))
-		Expect(index).To(Equal(0))
-	})
-
-	It("makeDeviceName should fill target gap when mixed bus types share same prefix", func() {
-		By("Setting up prefixMap simulating: SCSI boot=sda, SATA cdrom=sdc, SATA cloudinit=sdd, sdb available")
-		prefixMap := map[string]deviceNamer{
-			"sd": {
-				existingNameMap: map[string]string{
-					"boot":      "sda",
-					"cdrom":     "sdc",
-					"cloudinit": "sdd",
-				},
-				usedDeviceMap: map[string]string{
-					"sda": "boot",
-					"sdc": "cdrom",
-					"sdd": "cloudinit",
-				},
-			},
-		}
-
-		By("Adding new SCSI disk - should get sdb (the available gap)")
-		res, index := makeDeviceName("newhotplug", v1.DiskBusSCSI, prefixMap)
-		Expect(res).To(Equal("sdb"), "new disk should fill the gap at sdb")
-		Expect(index).To(Equal(1))
-
-		By("Adding another SCSI disk - should get sde (next available after sdd)")
-		res, index = makeDeviceName("anotherhotplug", v1.DiskBusSCSI, prefixMap)
-		Expect(res).To(Equal("sde"))
-		Expect(index).To(Equal(4))
-	})
 })
 
 var _ = Describe("direct IO checker", func() {
@@ -4890,9 +4678,9 @@ var _ = Describe("Driver Cache and IO Settings", func() {
 })
 
 func diskToDiskXML(arch string, disk *v1.Disk) string {
-	devicePerBus := make(map[string]deviceNamer)
+	devicePerBus := make(map[string]storage.DeviceNamer)
 	libvirtDisk := &api.Disk{}
-	Expect(Convert_v1_Disk_To_api_Disk(&convertertypes.ConverterContext{Architecture: archconverter.NewConverter(arch), UseVirtioTransitional: false}, disk, libvirtDisk, devicePerBus, nil, make(map[string]v1.VolumeStatus))).To(Succeed())
+	Expect(storage.Convert_v1_Disk_To_api_Disk(&convertertypes.ConverterContext{Architecture: archconverter.NewConverter(arch), UseVirtioTransitional: false}, disk, libvirtDisk, devicePerBus, nil, make(map[string]v1.VolumeStatus))).To(Succeed())
 	data, err := xml.MarshalIndent(libvirtDisk, "", "  ")
 	Expect(err).ToNot(HaveOccurred())
 	return string(data)
