+++ pkg/virt-launcher/virtwrap/converter/storage/BUILD.bazel
@@ -3,6 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
+        "disk.go",
         "virtiofs.go",
         "volume.go",
     ],
@@ -14,26 +15,38 @@ go_library(
         "//pkg/container-disk:go_default_library",
         "//pkg/emptydisk:go_default_library",
         "//pkg/host-disk:go_default_library",
+        "//pkg/pointer:go_default_library",
+        "//pkg/safepath:go_default_library",
+        "//pkg/storage/reservation:go_default_library",
+        "//pkg/storage/types:go_default_library",
         "//pkg/storage/volumepath:go_default_library",
+        "//pkg/util:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
+        "//pkg/virt-launcher/virtwrap/converter/iothreads:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter/types:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter/virtio:go_default_library",
+        "//pkg/virt-launcher/virtwrap/device:go_default_library",
+        "//pkg/virt-launcher/virtwrap/disksource:go_default_library",
         "//pkg/virtiofs:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
+        "//staging/src/kubevirt.io/client-go/log:go_default_library",
+        "//vendor/golang.org/x/sys/unix:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
     srcs = [
+        "disk_test.go",
         "storage_suite_test.go",
         "virtiofs_test.go",
     ],
+    embed = [":go_default_library"],
     race = "on",
     deps = [
-        ":go_default_library",
         "//pkg/libvmi:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
+        "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
